+++ src/FalconEngine/Graphics/Renderer/Resources/IndexBuffer.cpp
@@ -6,7 +6,7 @@ namespace FalconEngine
 /************************************************************************/
 /* Constructors and Destructor                                          */
 /************************************************************************/
-IndexBuffer::IndexBuffer(int indexNum, IndexType indexType, BufferUsage usage) :
+IndexBuffer::IndexBuffer(size_t indexNum, IndexType indexType, BufferUsage usage) :
     // The index buffer doesn't need the vertex array information about index
     // channel dimension. So we pass 1 here.
     Buffer(indexNum, indexType == IndexType::UnsignedShort
